requirements.txt
bidict==0.21.2
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
typing-extensions==3.10.0.0
Wand==0.6.6
Werkzeug==2.0.1
zipp==3.4.1
